% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerStatistic-package.R
\docType{package}
\name{cerStatistic-package}
\alias{cerStatistic}
\alias{cerStatistic-package}
\title{cerStatistic: Statistical Analysis for CER}
\description{
Statistical Analysis for CER.
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
